+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 **v0.3.4 - 08-28-2026**
 ---
 
-- Fix post delete ghost messages on message list
+- Fixed post delete ghost messages on message list
 - New app icon with full hicolor size set + scalable SVG - [#395](https://github.com/hkdb/aerion/issues/395) - [#99](https://github.com/hkdb/aerion/issues/99) - (Special thanks to @gianlucamazza)
+- Fixed thread sender and preview in message list - [#169](https://github.com/hkdb/aerion/issues/169)
 
 
 **v0.3.3 - 08-06-2026**